+++ src/.gitignore
@@ -1,3 +1,4 @@
 *.out
 coded.txt
 input.txt
+*.gch
