+++ src/cat.h
@@ -1,0 +1,15 @@
+//cat.h
+//
+
+#ifndef __CAT_H__
+#define __CAT_H__
+
+#include <stdlib.h>
+
+__BEGIN_DECLS
+int	cat __P((const char *));
+__END_DECLS
+
+#endif /* !__CAT_H__ */
+
+/* vim: set ts=4 sw=4 noexpandtab tw=79: */
