+++ src/decode.c
@@ -1,0 +1,18 @@
+//decode.c
+//
+
+#include <stdio.h>
+
+#include "decode.h"
+#include "bitconv.h"
+
+uint8_t
+decode(
+	const char *restrict ifile)
+{
+	printf("%s\n", "decode does nothing as of yet");
+	return 0; //stand-in
+}
+
+/* vim: set ts=4 sw=4 noexpandtab tw=79: */
+
