+++ src/decode.h
@@ -1,0 +1,15 @@
+//decode.h
+//
+
+#ifndef __DECODE_H__
+#define __DECODE_H__
+
+#include <stdint.h>
+
+__BEGIN_DECLS
+uint8_t	decode __P((const char *restrict));
+__END_DECLS
+
+#endif /* !__DECODE_H__ */
+
+/* vim: set ts=4 sw=4 noexpandtab tw=79: */
